requirements.txt
fastapi==0.104.1
starlette==0.32.0
itsdangerous==2.1.2
langchain==0.0.340
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.5
tiktoken==0.4.0
faiss-cpu==1.7.4
uvicorn==0.24.0